requirements.txt
pandas
numpy
flask
mlflow
dvc
dvc-s3
python-dotenv
pymongo[srv]==3.6
pymongo
certifi
scikit-learn


## -e .